fix : 같은 index 일 경우 안나오는 버그 수정
